requirements.txt
Bottleneck==1.3.2
nltk==3.4.5
nose==1.3.7
notebook==6.0.3
numba==0.48.0
numexpr==2.7.1
numpy==1.18.1
numpydoc==0.9.2
opencv-contrib-python-headless==4.3.0.36
openpyxl==3.0.3
opt-einsum==3.2.1
packaging==20.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pkginfo==1.5.0.1
pluggy==0.13.1
ply==3.11
PyYAML==5.3
pyzmq==18.1.1
SQLAlchemy==1.3.13
statsmodels==0.11.0
streamlit==0.63.0
sympy==1.5.1
tqdm==4.42.1

